blog-snapshots-func/requirements.txt
azure-functions
azure-storage-blob
azure-identity
azure-storage-queue==2.1.0